LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-reduced-padding.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing <code>MacOSInlineMediaControls</code> with reduced padding.");

const mediaControls = new MacOSInlineMediaControls({ width: 300 });
mediaControls.layoutTraits = LayoutTraits.macOS | LayoutTraits.ReducedPadding;

shouldBe("mediaControls.leftContainer.leftMargin", "12");
shouldBe("mediaControls.leftContainer.rightMargin ", "16");
shouldBe("mediaControls.leftContainer.buttonMargin", "16");
shouldBe("mediaControls.rightContainer.leftMargin ", "0");
shouldBe("mediaControls.rightContainer.rightMargin", "12");
shouldBe("mediaControls.rightContainer.buttonMargin ", "16");
debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
